[Point Instancers] Fix for removing nested instancers.
The Point instancers removal code was making the assumption that all prototypes were Rprims.  This isn't true, as a prototype can be another instancer.
The bug was found, because the point instancer was trying to remove the instancer as an rprim, which silently failed, this left the child instancers prototypes in the render index,
which left some side effect.  The first observed point was trying to get the render tag for the rprims that should have been deleted.  As Usd didn't know about them, it failed safe
to the geometry render tag, which caused "render" geometry to be drawn, filling up GPU memory.

When populating a point instancer, we handle the situation by delegating the responsibility of populating the prototype to that prototypes adapter.

The same solution will work for removal to, where the prototype's adapter should perform the actual removal.

Changes:
- In _UnloadInstancer(), delegate responsibility for removing a prototype to that prototypes adapter.

(Internal change: 1847462)