[pxrUsdMayaGL] improve flexibility when using _GetWireframeColor()
This change modifies PxrMayaHdShapeAdapter::_GetWireframeColor() so that it
*always* returns a wireframe color (if the mayaWireColor parameter is not
nullptr), either based on a soft selection that includes the object, or
otherwise as determined by MHWRender::MGeometryUtilities::wireframeColor().

The return value is a suggestion as to whether or not to use the wireframe
color under "normal" usage, that is if the object and/or its component(s)
are involved in a selection, or if the displayStyle indicates that
wireframes are being drawn. Callers are free to use this suggestion or not
if they want to use other criteria for determining whether to use the
returned wireframe color.

This change also fixes an issue where bounding boxes in the legacy viewport
were not getting drawn with the same framebuffer gamma correction that is
applied when drawing via Hydra.

(Internal change: 1847947)