[pxrUsdMayaGL] fix for batch renderer crash when opening a new scene
This change fixes a crash that can occur with shapes that are imaged by the
batch renderer and drawn immediately when the scene is opened (e.g. proxy
shapes *without* an assembly parent). The batch renderer was listening for
kAfterOpen scene messages from Maya to know when to reset, but these could
potentially be delivered after the scene was fully opened and had had a
chance to complete an initial draw. In that case, the batch renderer was
being reset out from under an active scene, and subsequent draws or
selections would lead to a crash.

The only scene message option that we receive at exactly the right time
(after any existing scene has been closed but before the new scene has been
opened) is kBeforeFileRead, but that is also emitted when a file is
imported or referenced into the current scene. As a result, we need to make
sure we do *not* reset the batch renderer if that is the case.

This was not caught previously because the existing unit test used assembly
nodes, which must be loaded into the Collapsed representation in order to
create the proxy. New test cases were added that use proxy shapes directly.

(Internal change: 1848227)