Anchor asset paths in UsdStage::Flatten
Flattening a UsdStage now anchors asset paths in
metadata and attributes to ensure that resolving
these paths on the flattened result will yield the
same files as on the original stage. This information
may otherwise be lost in certain cases, e.g. a
file-relative asset path in a sublayer located in a
different directory from the flattened result.

This is consistent with UsdUtilsFlattenLayerStack,
which implements another form of flattening.

(Internal change: 1848982)