API schemas are now classified into applied and non-applied API schemas. The author of an API schema has to decide on the type of API schema at the time of its creation by setting customData['apiSchemaType'] in the schema definition (in the associated primSpec in the schema.usda file).
If an API schema only provides an interface to set certain core bits of metadata
(like UsdModelAPI, which sets model kind and UsdClipsAPI, which sets clips-
related metadata) or if the API schema can apply universally to any type of prim
or only to a known set of prim types and there is no use of recording the
application of the API schema, in such cases, it would be better to make it a non-
applied API schema. Examples of non-applied API schemas include UsdModelAPI,
UsdClipsAPI, UsdShadeConnectableAPI and UsdGeomPrimvarsAPI.

If there is a need to discover (or record) whether a prim contains or subscribes
to a given API schema, it would be advantageous to make the API schema an
"applied" schema. In general, API schemas that add one or more properties to a
prim should be tagged as applied API schemas. A public Apply() (or private _Apply())
method is generated for applied API schemas by usdGenSchema. An applied
API schema must be applied to a prim via a call to the generated Apply()
method, for the schema object to evaluate to true when converted to a bool
using the explicit bool conversion operator. This makes the Apply() method on
an applied API schema akin to the Define() method on typed schemas. Examples of
applied API schemas include  UsdCollectionAPI, UsdGeomModelAPI and UsdGeomMotionAPI.

An applied API schema can only inherit from another applied API schema or
from UsdAPISchemaBase directly. Similarly, a non-applied API schema can
only inherit from a non-applied API Schema or directly from UsdAPISchemaBase.
These restrictions are now imposed by usdSchemaGen.For this reason,
UsdGeomModelAPI, which is an applied API schema no longer inherits from the
non-applied UsdModelAPI and instead inherits from UsdAPISchemaBase directly.

Removed support to the "isMultipleApply" customData key in schema-gen and
replaced it with a token valued "apiSchemaType" key which can be one of
{"nonApplied", "singleApply", "multipleApply"}. A non-applied API schema can
be added to a schema.usda file by setting customData["apiSchemaType"]="nonApplied".

An applied API schema can be single-apply or multiple-apply. For a single-apply
API schema, which is the default type for an API schema,
customData["apiSchemaType"] must be left unspecified or set to "singleApply"
explicitly. For a multiple-apply API schema, customData["apiSchemaType"] must
be set to "multipleApply" explicitly. Note that a multiple-apply API schema may
not inherit from a single-apply API schema or vice versa.

Other details:
* Updated usdGenSchema so it doesn't generate an apply method for the base
API-schema class UsdAPISchemaBase.

* Changed the signature of the virtual _IsCompatible method on UsdSchemaBase
to not take the prim, since it can be acquired inside the method if desired.
Also made the method protected, so that derived classes can invoke the base-
class implementation in addition to doing their own compatibility checks.

* usdGenSchema now lists the names of appliedAPISchemas in every generatedSchema.usda
file. This information is accumulated the schema registry which can now answer the
question of whether a given API schema is applied or non-applied.

* UsdAPISchemaBase's _IsCompatible() implementation now calls HasAPI() with the
(virtual) schema type to determine compatibility if the schema type represents an
applied API schema. i.e. a prim is compatible with an applied API schema type only
if it has been successfully applied to the prim.
An API schema class may choose to override this behavior if desirable, by not calling
the base implementation.

* Replaced the deprecated UnspecifiedBoolType conversion operator on
UsdSchemaBase with an implementation of the explicit bool operator.

* Removed the explicit bool conversion operator from UsdCollectionAPI and
added an _IsCompatible override in its place.

* Tagged UsdGeomPrimvarsAPI as a non-applied API schema.

* Updated many client sites where API schema objects were being checked
for validity using bool-conversion. These sites have been updated to either
check for the validity of the prim (where the schema isn't expected to be
applied), or by adding calls to Apply().

(Internal change: 1849919)
(Internal change: 1850071)
(Internal change: 1850099)
(Internal change: 1850176)
(Internal change: 1850401)