[katana] Remove resolvePaths argument
All asset paths are converted to the resolved path. Previously this was not
done when importing LightFilter and AttrMap. I believe this change is
necessary as not enough context is preserved to allow Renderman to deduce the
correct resolved path.

Removing this argument also allowed removal of the unused pathsAsModel argument.

Signed-off-by: Bill Spitzak <bill.spitzak@dreamworks.com>