[katana] Fixed conversion of VtArray<SdfAssetPath>
If asShaderParam was on, it returned a value as though it was off.
If asShaderParam was off, it returned an empty attribute.

This changes the code so the returned attribute is the same style as other
types (array of string if asShaderParam is on, type+array pair if
asShaderParam is off). Also a few formatting fixes to the cases so they all
match each other better.

Pass asShaderParam to ConvertRelTargetsToKatAttr. This seems consistent
and previous code may have been a workaround for the above behavior.

This changes behavior with importing of riStatements! Please check!

Signed-off-by: Bill Spitzak <bill.spitzak@dreamworks.com>