[usdMaya] modify stage cache clearing callback to mimic Maya/Hd batch renderer reset
This change applies a similar pattern of Maya scene message callbacks to
clearing the UsdMayaStageCache as is being used to reset the Maya/Hydra
batch renderer. This addresses an issue where the stage cache would be
cleared during a "Save As" operation, and it ensures that the stage cache
is cleared in between an existing scene being closed and a new scene being
opened.

(Internal change: 1853917)