[HdSt] Support VeryLow complexity
Replaces delegate GetRefineLevel with a GetDisplayStyle entry for querying
if the application wants to disable smooth normals or displacement as well
for a prim.

HdStMesh uses the new struct to determine if either should be disabled.
The change also rolls up existing topology checks into the same booleans.

(Internal change: 1854009)