[Hd] Add HdRenderThread utility class, to enable render delegates to render in a background thread.
HdRenderThread provides a centralized utility class for a render delegate to manage synchronization with a background thread.  It has a relatively constrained API and callback interface, so the hope is that render delegate authors can use it to enable background rendering while largely avoiding the pitfalls that asynchronous computation inevitably introduces.  The hope is also that by providing a single recommended threading model, we concentrate all the documentation and bugfixes in one place.

(Internal change: 1854039)