[usdMaya] make the unassigned value for uv's (0.0, 0.0) instead of a large negative number.
Originally, our reason for using -1e30 was that since it's unassigned we
wanted to make it look as intentionally unassigned as possible.

Any downstream client that is handling unassigned values correctly will
not care what this value is.  But some clients that do not handle it
correctly would prefer a value more natural default.

(Internal change: 1854669)