Fix usdImaging generation of texture IDs.
Hydra texture IDs are keys into the texture hash, so it's important for them to be unique.  Previously, usdImaging computed hashes by taking the usd path hash of the texture prim.  This has two issues:
- usd paths, being SdfPath, base their hashes on internal pointer values instead of hashed contents, and so they're not good for long-term use.
- with variants, the same usd path might point to a totally different asset path.

Now, we hash on the asset path (which is a proper hash-on-contents), and also metadata.  This may result in deduplication of textures if a usd file comes in with lots of duplicated texture references.

(Internal change: 1854784)