[Pcp] Remove use of std::auto_ptr.
This is deprecated in C++11, and removed in C++17, so we shouldn't rely
on it. Also, auto_ptr was, and always has been, a not-so-great idea.
Note that this uses shared_ptr, we could use unique_ptr if we relied on
a newer version of boost, as far as I know.