Fix missing resync when adding an inert spec
Fixes a bug where adding an inert spec would not be
considered a significant resync if the affected prim
index had previously been computed but had no specs.

This was the root cause of an issue in Usd where
adding a new override, removing it, then adding it
again would return an invaild prim.

(Internal change: 1855843)