Usd: usdGenSchema now allows setting apiName customData to an empty string to request that no Get/Create API be generated.
An example where this is useful is pre-applying an API schema
such as UsdCollectionAPI; we don't want to emit API for
the implementation attributes of the schema.

Note: whitespace in wrapSchemaClass.cpp is there to preserve
generated whitespace to minimize diffs.  Previously there was
trailing whitespace on a preceding line, which is meaningful
to Jinja; putting it on its own line hopefully makes its
presence more obvious.

(Internal change: 1856016)