Remove textReferenceParser from pxr
SdfExtractExternalReferences was used to find
references to external assets in ASCII layers by
scanning the file for the syntax used by the
text file format to write out asset paths. This
was a maintenance burden, as it required
making sure it was kept up-to-date with any
changes to the ASCII format. It also introduced
a dependency on boost::regex(*), which we'd
like to remove.

So, this code is being removed from pxr. The only
client of this was UsdUtilsExtractExternalReferences
API; the results from this function should be
unchanged, except for differences in ordering.

(*) Ideally, this code would use std::regex instead,
but the std::regex implementation is incomplete on
some versions of libstdc++.

(Internal change: 1856710)
(Internal change: 1856921)