UsdCollectionAPI can now represent collections that include everything, as well as collections that exclude certain paths but include all others.  Relationship targets cannot target the root path </>, so we instead encode this as a boolean attribute "includeRoot", scoped to the collection.  For consistency with existing assets, its fallback value is false.
UsdCollectionAPI::MembershipQuery can now also return the
path-to-expansionRule map, for DCC plugins that want to
efficently translate the rules into native form.

The motivation is to enable use of core USD collections for
linking in UsdLux, replacing UsdLuxLinkingAPI.	By default,
lights link to all geometry; it is also possible for them to
link everything but a few objects (for example, to partition a
set of lights just for character eyes).

(Internal change: 1857141)