Add ArchAbsPath and ArchNormPath
Move the implementations of TfAbsPath and TfNormPath
into arch. This follows our convention that code
that is architecture-dependent should live in arch.
It also makes this functionality available to code
in arch, which will be useful in an upcoming change.

Note there are slight modifications to the code to use
std::replace instead of TfStringReplace and std::string's
cbegin() instead of a custom helper function.

(Internal change: 1857747)
(Internal change: 1857767)
(Internal change: 1857783)