Fix tests for static USD builds
With change 1857400, executables that statically link
against USD libraries will now look for resource files
in a "usd" directory relative to where the executable
is installed. This caused tests to fail, since these
tests are installed in a different directory and do
not include a copy of these resource files.

Previously, these tests worked because they had
been looking for these files in "../share/usd/plugins",
which just happened to be the right place because
of where tests are installed.

Instead of copying these resource files with the
tests, we now set the PXR_PLUGINPATH_NAME
environment variable to point to the installed
resource files when running tests from a static
build.

(Internal change: 1857787)