Try to address the TLS 1.2 requirement on Windows
When the sites where build_usd.py gets the dependencies switched to
requiring TLS 1.2, we updated the build_usd.py script to look for curl
and use that where it can find it.  The idea being that it would
avoid problems where too old a version of Python is installed, and urllib2
wouldn't have the proper support.

Unfortunately for Windows machines, curl is not typically there.  So now
we're trying to use powershell in those cases.

Fixes issue #449

(Internal change: 1858046)