[UsdMaya] Deal with instancer data that has no ids array; compute number of instances from largest data array
Previously, we assumed that all instancers' input data would contain an ids array, and we used this ids array to
determine how many instances there were. It turns out that a dynamics setup with MASH produces data without
an ids array. Since ids isn't required in UsdGeomPointInstancer, we just skip it in that case. The number of instances
is now computed by taking the size of the largest instancer data array (since some may be sparse).

(Internal change: 1858818)