This change enhances the ways in which a UsdShadeShader can identify its shader program.
Added a built-in, token-valued "info:implementationSource" attribute to
UsdShadeShader. This can be set to one of {"id", "asset", "code"}.
  - "id" indicates that the "info:id" attribute should be consulted to get the
  source shader program (using the shader registry).
  - "asset" indicates that the "info:{sourceType}:sourceAsset" attribute must
  be read to get the path to the asset containing the shader program.
  - "code" indicates that the shader program is available inline as the value
  of the attribute "info:{sourceType}:sourceCode".

Added convenience API to Get/Set ID, sourceAsset or sourceCode on a shader.

{sourceType} above is a place holder for a token that identifies the type of
shader source, for example: oso, osl, glslfx, riCpp etc.
When sourceType is empty, the corresponding source asset or the source code
is considered to be "universal" (or fallback). When the source asset (or code)
corresponding to a specific sourceType is unavailable, then the universal
source asset (or code) is returned by the GetSource{Asset,Code} API, if present.

The fallback value of the "info:implementationSource" attribute is "id", which
preserves the existing behavior.

(Internal change: 1859133)