[Hd] Added support for masked selection highlighting in Hydra
In Hydra, traditional selection highlighting of prims, instances and subprims is done in the fragment shader via a global selection buffer, before lighting/shading.
Masked selection highlighting differs from it, in that, vertex weights are used in conjuction with a mask color, and applied after lighting/shading.

ApplyPostLightingColorOverrides has been added as a post-step in the shading terminal, which is implemented in various mixins in {hdSt, hdx}/renderPass.glslfx.

(Internal change: 1859340)