Hd, UsdImaging: Light linking.
In USD, lights use collections to indicate which objects
they illuminate and which objects cast shadows.  This
resembles how linking is expressed in an app like Katana.

In a renderer, this representation is often renormalized
to better service the queries it makes while computing light
transport.  A renderer shading a point needs to query which
lights contribute direct illumination; a light shader computing
that direct illumination needs to query which objects may
occlude that path.

Renderers also prefer to cluster participants in these
relationships, to track them with the minimum number
of unique rows in their lighting matrix.  In Renderman
this is done using the grouping:membership attribute.
Mantra has a similar concept known as categories.
Both assign geometry and lights to named groups and
refer to those groups to establish linking.

We follow suit in Hydra.  We add a new GetCategories() API,
returning a VtArray of TfTokens.  We choose the term "categories"
here since it seems less likely to cause confusion than "group",
which is more often used to indicate hierarchical groups in
the scene.

Light- and shadow-linking are represented as new light parameters,
HdTokens->lightLink and shadowLink.  For now, they are a TfToken
that refers to a single category.

UsdImaging consumes the collections on UsdLuxLight
prims.  It tracks all lighting collections encountered
in a CollectionCache.  It tracks them by their
computed MembershipQuery, so matching collections
fall into an equivalence class.  Each unique
query is assigned a name; for debuggability the
path of the first collection encountered is used.

Future work / open questions

* Unit test coverage
* Invalidation and change processing needed
* ComputeCollectionsContainingPath() loops over
  all unique collection queries; in the presence
  of many queries we may need a hierarchical
  organization of the queries, for faster rejection.
* Some renderers support more elaborate category
  syntax for linking; do we want to support
  linking expressions, or only a baked-down form?

(Internal change: 1859390)