Allow empty LibraryPath in plugInfo.json
The LibraryPath entry in a plugInfo.json file points to
the library that should be loaded during a call to
PlugPlugin::Load. However, this entry is superfluous
and misleading when dealing with monolithic USD
libraries or individual static libraries; in those cases,
the various USD libraries are not separately loadable.

With this change, the build will now fill in LibraryPath
with an empty string for the above build configurations,
and PlugPlugin will treat this as a valid entry.

Aside from fixing the conceptual issue above, this
allows users performing a monolithic build to use the
same set of plugInfo.json files for both the shared
usd_ms and and static usd_m libraries.

(Internal change: 1859456)