Fix value clip time mapping bug
Fixes an issue where clip time mappings would
not be taken into account if the mapping's stage
time fell outside the time when a given clip was
active.

For example, if clip 1 was active in the time range
[0.0, 2.0) and the stage-to-clip time mappings were
[(0.0, 0.0), (3.0, 3.0)], the second mapping would
be ignored. This would cause requests for the value
at time 1.0 to use the value in clip 1 at time 0.0
instead of 1.0 (which we arrive at by interpolating
between the two mappings).

(Internal change: 1860137)
(Internal change: 1860140)