Hd,UsdImaging: UsdImaging no longer uses UsdRi.  To look up material networks it now uses UsdShadeMaterialAPI.
To suport this, HdRendererDelegate now provides a material
selector token to indicate which (of possibly several) material
network implementations to use.  In UsdShadeMaterialAPI this is
known as the "renderer context" but we felt that is not a
specific enough term in a rendering subsystem like Hydra,
so we opted for "material network selector" instead.

(Internal change: 1863051)