Build Katana plugins on Windows.
One thing to shout out here is the deliberate change in /cmake/macros/Private.cmake:

     set(apiPublic "")
     set(apiPrivate ${uppercaseName}_EXPORTS=1)
     if(NOT _building_monolithic AND args_TYPE STREQUAL "STATIC")
-        set(apiPublic PXR_STATIC=1)
+    #    set(apiPublic PXR_STATIC=1)
     endif()

I've had to do this because there are 3 static libraries that the usdKatana target links against. The above PXR_STATIC compile definition is a public property on each of the 3 static libraries and is therefore propagated to usdKatana, which is a dynamic library.

This causes usdKatana to attempt to statically link all of its link dependencies rather than import.

Whilst my above "fix" works for us, it shouldn't make its way into the public repo.