[pxrUsdMayaGL] fix for deleted shapes continuing to draw in the legacy viewport
In Viewport 2.0, the MPxDrawOverride destructor is called when its shape is
deleted, in which case the shape's adapter is removed from the batch
renderer. In the legacy viewport though, that's not the case. The
MPxSurfaceShapeUI destructor may not get called until the scene is closed
or Maya exits. As a result, MPxSurfaceShapeUI objects must listen for node
removal messages from Maya and remove their shape adapter from the batch
renderer if their node is the one being removed. Otherwise, deleted shapes
may still be drawn.

(Internal change: 1863388)