[UsdMaya] Add a batching diagnostic delegate for Maya.
This delegate converts any Tf statuses/warnings/errors into Maya infos/warnings/errors that appear in the Maya script window.
The batching functionality is invoked by constructing a PxrUsdMayaDiagnosticBatchContext in C++ or using the context
manager-style object UsdMaya.DiagnosticBatchContext in Python. As long as some batch context remains alive, warnings/statuses
are batched, and as soon as the last one is destroyed, the warnings/statuses are coalesced and printed
(using UsdUtilsCoalescingDiagnosticDelegate). Errors are always printed immediately; this matches the behavior in
UsdUtilsCoalescingDiagnosticDelegate.

The delegate is installed by pxrUsd when the plugin is loaded, and it is removed when the plugin is unloaded from Maya.
The batch context deals with cases where the delegate is not available, or is removed during the middle of a batch context.

(Internal change: 1865098)