[HdSt] Enhance the current implementation of batching in HdSt
With the current aggregation strategy, there is a chance to end up with one batch per draw item, which could affect rendering performance.
For example, if you have a scene with many draw items, the draw items could map onto the same batchMap key which alternate into multiple different aggregated buffers.
This PR provides managing multiple aggregated buffers for each batchMap key to reduce the number of draw batches.

The test scene has
 - rprims: 8029
 - instancer: 7429 (1 prototype for each instancer)
 - instances: 28132
 - textures: 231
and everything is static with an animated camera.

The current implementation generated
 - 156 draw batches,
 - 8029 draw items
and it attempted to append a batch 3 times.

This PR generated
 - 7 draw batches,
 - still 8029 draw items
and it attempted to append a batch 11,319 times.
The batchMap had 3 keys and the depth of each key's vector was 1, 5, and 1.
The test scene was able to be rendered upto 10fps performance improvement.