[HdSt] Add Hydra Stream support for limiting Texture Memory.
The Glf texture system, gives users the opportunity to limit the amount of memory used by only loading mip levels up to that memory limit.

The system also supports share the same texture, therefore a specific texture can have multiple request, in which case the policy is the largest request wins.

However this is where the complexity comes in, as we need to manage the lifespan and remove the request when the texture isn't used anymore.

Therefore, the management of the request has been added to HdStTextureResource.  The memory request is added in the constructor and removed in the destructor of the object.

The alternative constructor was used by the test framework and to create a fallback texture, so rather than add memoryRequest support to that constructor, the code was switched to all use the same constructor.

Changes:
- In GetTextureResource() switch the wrap and filter modes for the fallback texture to be more appropriate/efficient.
- In HdStMaterial, switch to the fully specified constructor for the fallback texture.
- In HdTexture, trigger garbage collection when the texture resource is changed, so the texture cache doesn't hold on to a stale memory request.
- The HdStSimpleTextureResource() constructor now takes a memory request param and adds it to the texture handle.  The request is removed in the destructor.
- HdSt_UnitTestDelegate now uses the fully specified constructor, using the same values for wrap modes and filtering.

(Internal change: 1865306)