[UsdMaya] Don't CHECK_MSTATUS_AND_RETURN_IT when stage can't be loaded by reference assembly
This seems to preserve the same behavior evaluation-wise, but (1) makes errors appear
in UI instead of in stderr, and (2) ignores errors when the file path is empty, so that
initially creating an empty reference assembly doesn't throw an error.

This is a user-facing problem, and not an internal check, so use TF_RUNTIME_ERROR
to display an appropriate user message in UI. Furthermore, we are still able to provide
the appropriate Maya data if the stage isn't loaded. (Silently return an empty data
object if the stage can't be loaded because the file path isn't specified.)

(Also removes a "return MS::kFailure" that didn't do anything in the proxy shape)

(Internal change: 1866231)