[usdMaya] add pxrUsdStageNode for reading a USD stage and providing it to downstream Maya nodes
This change adds a simple MPxNode that reads in the USD stage identified by
its file path attribute and makes that stage available as a stage data
object on its output attribute. Downstream Maya nodes can connect this
output to their own stage data input attributes to gain access to the
stage. This allows sharing of a single USD stage by multiple downstream
consumer nodes, and it keeps all of the specifics of reading/caching USD
stages and layers in this stage node so that consumers can simply focus on
working with the stage and its contents.

(Internal change: 1867288)