[usdMaya] add pxrUsdPointBasedDeformerNode that deforms Maya geometry using a UsdGeomPointBased prim
This Maya deformer node can be used to deform Maya geometry to match the
points of a UsdGeomPointBased prim. It takes as input a pxrUsdStageData
object (which can be received from a connection to a pxrUsdStageNode), the
prim path to a UsdGeomPointBased prim in the stage data's stage, and a time
sample. When the deformer runs, it will read the points attribute of the
prim at that time sample and use the positions to modify the positions of
the geometry being deformed.

(Internal change: 1867444)