[HdSt] Selecting face 0 no longer causes all points to be highlighted. Points is its own repr (using HdGeomstylePoints and a custom terminal), and internally builds a topology buffer without the primitiveParam binding. While adding support for custom point highlighting, a subtle bug was introduced in that all points are highlighted when face 0 is selected. This stems from the fallback code generated for points, wherein GetElementID returns 0, which is in turn used during subprim highlighting. The points end up being highlighted when face 0 is selected because the fragments claim that they're from a selected face.
This CL fixes that, and ensures element and fvar primvars are skipped when rendering a mesh as points (note: only a mesh, since we don't yet support point picking on basisCurves). This allows us to use a fallback GetElementId that returns -1 safely, since there isn't any caller using it as an index (which would have resulted a runtime error).

(Internal change: 1867679)