Fix change processing bug with nested instances
In cases where a UsdStage had multiple instances of
an asset that itself had multiple instances of
another asset, loading and unloading the top-level
instances repeatedly could lead to crashes. This
was because UsdStage would select the wrong source
prim index for the inner instances, which caused
the master prim for these instances to not be
generated.

(Internal change: 1867691)