[HdSt] If mesh vertex primvar data is larger than expected, truncate it to the expected length.
We allocate vertex primvar ranges based on the size of the first vertex buffer source.  This means when you pass in over-sized primvar data, it's inconsistent whether you over-allocate the range, or truncate at buffer upload time.

With this change, the primvar data is always truncated, at the site of the validation warning about oversize primvar data.

(Internal change: 1867729)