Updates to Ndr and Sdr to handle creation and fetching of shader nodes from sourceAsset and sourceCode values on a UsdShadeShader.
Added an optional string-valued "sourceCode" data member to NdrNodeDiscoveryResult.
When resolvedUri field of a discovery result is empty, parser plugins must consult the
sourceCode string to see if the source code for the node is available in-memory. When
the "sourceCode" field in a DiscoveryResult object is non-empty, it is parsed by parser
plugins to determine inputs, outputs, metadata etc. It is also copied into the NdrNode, for
use by the renderer.

Added API on NdrRegistry to create and add a node to the registry from a given asset path
(GetNodeFromAsset) or from a given sourceCode string (GetNodeFromSourceCode). This API
takes an additional dictionary valued "metadata" argument to allow clients to specify
additional metadata (eg, primvar names) that may be needed to process the source asset
or source code approriately. A unique identifier is constructed based on a hash of the asset
path value or the sourceCode string and the metadata dictionary in each case. This becomes
the identifier and the name of the NdrNode that is created by these methods.

(Internal change: 1867834)
(Internal change: 1867881)