Updates to UsdShade in support of creation and fetching of shader nodes for Shaders whose implementation source is either "sourceAsset" or "sourceCode".
* Added a dictionary valued metadata field named "shaderMetadata" on UsdPrims. Since the field only applies
to shader prims, the supporing API is implemented in UsdShadeShader. This field is used to specify additional
metadata (eg. primvar names) that may be needed to process the source code of the shader correctly.
* Added API to create (if necessary) and fetch the SdrShaderNode corresponding to a Shader prim, for a specified
sourceType, from the shader registry (SdrRegistry). A new node will be created in the registry if the shader's
implementation source is "sourceAsset" or "sourceCode" and the corresponding info:sourceAsset or info:sourceCode
value is being encountered for the first time. UsdImagingMaterialAdapter will use this API in an upcoming change
to populate the 'type' field (which should really be renamed to "identifier") of the HdMaterialNode it constucts to
pass on to the render delegate. The render delegate can lookup the node in the registry by identifier.

(Internal change: 1867835)
(Internal change: 1867882)