[UsdMaya] Combine MayaPrimWriter::ExportsReferences() and MayaPrimWriter::GetAllAuthoredUsdPaths() into a single MayaPrimWriter::GetModelPaths()
MayaPrimWriter::GetAllAuthoredUsdPaths() was badly named; it was really intended as a way to tell
the PxrUsdMaya_ModelKindWriter which prim paths were models and needed to have kind fixed up
along their hierarchy. To further confuse things, GetAllAuthoredUsdPaths() was only requested
whenever ExportsReferences() was true. (Note that the confusing naming led to GetAllAuthoredUsdPaths()
being defined in prim writers where it was never used, e.g. in the skeleton writer.)
This combines their behavior, so that the new GetModelPaths() explicitly tells the model
kind writer which paths to check for kind. The default preserves behavior by returning an empty vector for
GetModelPaths(), which does the same thing as ExportsReferences() = false.

(Internal change: 1871202)