[HdSt/Glf] Improved Texture Error handling.
As textures are assets, there are many reasons why a texture could result in an error.  The system wasn't quite robust in the event of an error, which lead to a user reporting a crash.

It is desirable behavior, that if there is an issue with the texture binding in the material, such as it is unbound or there is a problem with the texture asset, then the "fallback" value associated with the texture binding should be used.  In order for this to work, the error's with the texture need to be propagated to the material, so the texture system needs to be able to operate in an error state.

Changes:
- It is now a coding error, to pass a null texture when creating a texture handle.  The was resulting in a null ptr dereference in GlfTextureHandle::_ComputeMemoryRequirement(), which was the original crash in the bug.
- GlfTextureRegistry::GetTextureHandle(GlfTextureRefPtr) now returns a null texture handle if the ref ptr is null.
- HdStSimpleTextureResource no longer requires the texture handle to be non-null.  Fixed some dynamic casts that were not checking for a null result.
- Refactored HdStMaterial::_GetTextureResource(), so that if any error occurs trying to get the texture resource the fallback texture is created and used (previously, the fallback texture was only used for an invalid texture id).

(Internal change: 1871575)