[usdMaya] add a 'useAsAnimationCache' parameter to usdImport
When importing animated USD with this parameter enabled, usdImport will
create a pxrUsdStageNode for the USD file that is being imported. Then for
each geometry prim being imported that has time-sampled points, a
pxrUsdPointBasedDeformerNode will be created that reads the points from USD
and uses them to deform the imported Maya geometry. This should provide
much better import and playback performance when importing time-sampled
geometry from USD, and it should reduce the weight of the resulting Maya
scene since it will bypass creating blend shape deformers with per-object,
per-time sample geometry.

Note that this means that a link is created between the resulting Maya
scene and the USD file that was imported. With this parameter off, as is
the default, the previous blend shape deformer behavior is preserved, and
the USD file that was imported can be freely changed or deleted post-import.
With the parameter on, however, the Maya scene will have a dependency on
that USD file, as well as other layers that it may reference.

In this first implementation, only Mesh prims/Maya mesh nodes are
supported, but this should be fairly easily extensible to other point-based
prim/node types.

(Internal change: 1871925)