[UsdMaya] Support for exporting materials on prims in nested instances via collection-based bindings
Warnings are generated when instanced materials cannot be exported because collection-based bindings
are off. I've left an XXX to indicate a part of the code where we could be smarter to have _some_ instanced
bindings work without collection-based bindings, but since collection-based bindings will work for
everything, that seems largely unneeded for the moment.

Some non-nested instanced prims' materials now only export properly when collection-based bindings are
enabled due to the way that instance paths are computed for the mesh writer plugins.
(This isn't currently a big deal because instanced material support is relatively new; it is solvable if we wanted to tackle it -- see above.)
Thus testUsdExportShadingInstanced has been modified to export with collection-based bindings.

(Internal change: 1872177)